Factored Cocoa display utils into standalone module.
