+++ src/cocoa/toga_cocoa/widgets/tree.py
@@ -3,69 +3,7 @@ from rubicon.objc import *
 from ..libs import *
 from ..utils import process_callback
 from .base import Widget
-
-
-class TogaNodeData(NSObject):
-    @objc_method
-    def copyWithZone_(self):
-        copy = TogaNodeData.alloc().init()
-        copy.node = self.node
-        return copy
-
-
-class TogaIconCell(NSTextFieldCell):
-    @objc_method
-    def drawWithFrame_inView_(self, cellFrame: NSRect, view) -> None:
-        # The data to display
-        label = self.objectValue.node.data[0]
-        icon = self.objectValue.node.icon
-        if icon:
-            nicon = icon._impl.native
-            offset = 18
-
-            NSGraphicsContext.currentContext.saveGraphicsState()
-            yOffset = cellFrame.origin.y
-            if view.isFlipped:
-                xform = NSAffineTransform.transform()
-                xform.translateXBy(0.0, yBy=cellFrame.size.height)
-                xform.scaleXBy(1.0, yBy=-1.0)
-                xform.concat()
-                yOffset = 0 - cellFrame.origin.y;
-
-            interpolation = NSGraphicsContext.currentContext.imageInterpolation
-            NSGraphicsContext.currentContext.imageInterpolation = NSImageInterpolationHigh
-
-            nicon.drawInRect(
-                NSRect(NSPoint(cellFrame.origin.x, yOffset), NSSize(16.0, 16.0)),
-                fromRect=NSRect(NSPoint(0, 0), NSSize(nicon.size.width, nicon.size.height)),
-                operation=NSCompositingOperationSourceOver,
-                fraction=1.0
-            )
-
-            NSGraphicsContext.currentContext.imageInterpolation = interpolation
-            NSGraphicsContext.currentContext.restoreGraphicsState()
-        else:
-            # No icon; just the text label
-            offset = 0
-
-        if label:
-            # Find the right color for the text
-            if self.isHighlighted():
-                primaryColor = NSColor.alternateSelectedControlTextColor
-            else:
-                if False:
-                    primaryColor = NSColor.disabledControlTextColor
-                else:
-                    primaryColor = NSColor.textColor
-
-            textAttributes = NSMutableDictionary.alloc().init()
-            textAttributes[NSForegroundColorAttributeName] = primaryColor
-            textAttributes[NSFontAttributeName] = NSFont.systemFontOfSize(13)
-
-            at(label).drawAtPoint(
-                NSPoint(cellFrame.origin.x + offset, cellFrame.origin.y),
-                withAttributes=textAttributes
-            )
+from .utils import TogaIconCell, TogaNodeData
 
 
 class TogaTree(NSOutlineView):
@@ -188,9 +126,6 @@ class Tree(Widget):
             self.tree.expandItem(node._impl)
         else:
             self.tree.collapseItem(node._impl)
-
-        # node._impl = TogaNodeData.alloc().init()
-        # node._impl.node = node
 
     def refresh(self):
         self.tree.reloadData()
